Implement auth when deleting posts

+++ backend/src/apps/posts/domain/use-cases/remove-post.js
@@ -2,7 +2,7 @@ import makePost from "../entities/post/index.js";
 export default function makeRemovePost({ postsDb }) {
   return async function removePost(postDetails) {
     const post = makePost(postDetails);
-    let { error } = await postsDb.remove(post.getId());
+    let { error } = await postsDb.remove(post.getId(), post.getUserId());
     if (error) {
       throw new Error(
         `Error deleting post from database: ${error.message}. Post deletion failed.`
